Put calculator keys in another component
